requirements.txt
aenum==3.1.15
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
decorator==5.1.1
deepgram-sdk==3.7.3
deprecation==2.1.0
diffusers==0.31.0
executing==2.1.0
fastapi==0.115.2
ffmpy==0.4.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai @ git+https://github.com/google-gemini/generative-ai-python@e09b902274ebf405c55f0b3bd5b6affd61e5de02
googleapis-common-protos==1.65.0
gradio==5.1.0
gradio_client==1.4.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.5.0
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.2
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rsa==4.9
ruff==0.6.9
safetensors==0.4.5
selenium==4.25.0
semantic-version==2.10.0
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.40.0
sympy==1.13.1
tokenizers==0.20.1
tomlkit==0.12.0
torch==2.5.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.0
trio==0.27.0
trio-websocket==0.11.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
Wave==0.0.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
XlsxWriter==3.2.0
yarl==1.15.4
zipp==3.20.2
